Fix Assimp's CMakeLists.txt so the lib built successfully on Xcode. Close #709. For Travis CI - when performing a release, keep the CI mirror branches around for a little longer just in case we need to restart the CI job again due to Travis CI build error.

+++ Rakefile
@@ -281,8 +281,8 @@ end
 # Usage: NOT intended to be used manually
 desc 'Delete CI mirror branch'
 task :ci_delete_mirror do
-  # Skip if there are more commits since this one
-  if `git fetch -qf origin master; git log -1 --pretty=format:'%H' FETCH_HEAD` == `git show -s --format='%H' #{ENV['TRAVIS_COMMIT']}`.chomp
+  # Skip if there are more commits since this one or if this is a release build
+  if `git fetch -qf origin master; git log -1 --pretty=format:'%H' FETCH_HEAD` == `git show -s --format='%H' #{ENV['TRAVIS_COMMIT']}`.chomp && !ENV['RELEASE_TAG']
     system 'git config user.name $GIT_NAME && git config user.email $GIT_EMAIL && git remote set-url --push origin https://$GH_TOKEN@github.com/$TRAVIS_REPO_SLUG.git'
     system "git push -qf origin --delete #{ENV['TRAVIS_BRANCH']}" or abort "Failed to delete #{ENV['TRAVIS_BRANCH']} mirror branch"
   end
